requirements.txt
bcrypt==4.0.1
cryptography
fastapi
psycopg2-binary
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
httpx
pytest-asyncio
aiofiles
python-dotenv